requirements.txt
appnope==0.1.0
asgiref==3.2.10
authenticator==1.1.3
autopep8==1.5.4
coverage==5.2.1
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
coverage==5.2.1
cryptography==3.1
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
django-extensions==3.0.8
django-filter==2.3.0
django-seed==0.2.2
djangorestframework==3.11.1
factory-boy==3.0.1
Faker==4.1.2
gunicorn==20.0.4
ibm-cloud-sdk-core==1.7.3
ibm-watson==4.7.1
idna==2.10
inflection==0.5.1
ipdb==0.13.3
ipython==7.18.1
ipython-genutils==0.2.0
iso8601==0.1.13
jedi==0.17.2
Markdown==3.2.2
mock==4.0.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
traitlets==5.0.4
urllib3==1.25.10
Utils-Config==1.0.0
watson-developer-cloud==2.10.1
wcwidth==0.2.5
websocket-client==0.48.0
